Compute grouped summary statistics for a community

Calculate summary statistics (total, mean, and standard deviation for abundance, biomass, and metabolic rate) for a community, with flexible grouping options.

community_summarize(
  community,
  level = c("year", "species", "species_and_year", "custom"),
  id_vars = NULL
)

Arguments

community

a dataframe with one row per individual, columns for individual_mass and individual_bmr, and any additional relevant grouping columns. For example, the output of community_generate.

level

a character string specifying whether to group according to "year", "species", "species_and_year", or "custom". To use "custom", specify the grouping variables using id_vars

id_vars

a vector of column names to group by (if using level = "custom")

Value

a dataframe with the total, mean, and standard deviation of body mass and metabolic rate summarized over the individuals in community, grouped according to the grouping variables.

Details

For a dataframe of individuals from either numerous populations (of different species, or of the same species at different points in time), calculate summary statistics for each population or timestep, or for a different set of custom-specified grouping variables.

Examples

bbs_route <- filter_bbs_survey(demo_route_raw)
community_data <- community_generate(bbs_route)
community_summarize(community_data)
#>        routedataid countrynum statenum route rpid year species_designator
#> 1    9009911011994        900       99     1  101 1994                AOU
#> 63   9009911011995        900       99     1  101 1995                AOU
#> 116  9009911011996        900       99     1  101 1996                AOU
#> 172  9009911011997        900       99     1  101 1997                AOU
#> 231  9009911011998        900       99     1  101 1998                AOU
#> 274  9009911011999        900       99     1  101 1999                AOU
#> 318  9009911012000        900       99     1  101 2000                AOU
#> 366  9009911012001        900       99     1  101 2001                AOU
#> 423  9009911012002        900       99     1  101 2002                AOU
#> 481  9009911012003        900       99     1  101 2003                AOU
#> 547  9009911012004        900       99     1  101 2004                AOU
#> 607  9009911012005        900       99     1  101 2005                AOU
#> 683  9009911012006        900       99     1  101 2006                AOU
#> 734  9009911012007        900       99     1  101 2007                AOU
#> 789  9009911012008        900       99     1  101 2008                AOU
#> 833  9009911012009        900       99     1  101 2009                AOU
#> 888  9009911012010        900       99     1  101 2010                AOU
#> 933  9009911012011        900       99     1  101 2011                AOU
#> 975  9009911012012        900       99     1  101 2012                AOU
#> 1022 9009911012013        900       99     1  101 2013                AOU
#> 1074 9009911012014        900       99     1  101 2014                AOU
#> 1138 9009911012015        900       99     1  101 2015                AOU
#> 1184 9009911012016        900       99     1  101 2016                AOU
#> 1233 9009911012017        900       99     1  101 2017                AOU
#> 1280 9009911012018        900       99     1  101 2018                AOU
#> 1325 9009911012019        900       99     1  101 2019                AOU
#> 1377 9009911012020        900       99     1  101 2020                AOU
#> 1423 9009911012021        900       99     1  101 2021                AOU
#> 1494 9009911012022        900       99     1  101 2022                AOU
#>      total_abundance total_biomass total_metabolic_rate total_richness
#> 1               1361      156717.5             352041.7             27
#> 63              1443      162153.6             365924.9             27
#> 116             1413      167504.2             370367.6             27
#> 172             1381      159581.0             358457.0             27
#> 231             1415      156158.3             356129.4             27
#> 274             1412      164983.8             369193.8             27
#> 318             1423      156209.9             355571.4             27
#> 366             1462      168559.4             380097.7             27
#> 423             1443      170993.5             379842.9             27
#> 481             1431      169215.6             377787.3             27
#> 547             1484      176544.0             391518.8             27
#> 607             1499      167962.8             379294.2             27
#> 683             1450      168010.8             375160.3             27
#> 734             1450      175203.7             386228.1             27
#> 789             1449      163185.1             368108.1             27
#> 833             1403      152787.1             348133.9             27
#> 888             1419      154440.0             349824.4             27
#> 933             1488      182991.3             403678.3             27
#> 975             1470      167004.8             376819.0             27
#> 1022            1443      166238.8             373483.0             27
#> 1074            1439      154357.2             354965.6             27
#> 1138            1374      161747.4             361719.3             27
#> 1184            1442      156377.5             357137.5             27
#> 1233            1430      180636.2             396479.7             27
#> 1280            1415      152781.4             350457.8             27
#> 1325            1465      168744.5             380400.1             27
#> 1377            1441      168769.3             379576.0             27
#> 1423            1437      170192.0             378694.2             27
#> 1494            1453      171618.2             383058.4             27
#>      mean_individual_mass sd_individual_mass mean_metabolic_rate
#> 1                115.1488           177.8187            258.6640
#> 63               112.3726           175.2106            253.5862
#> 116              118.5451           188.1015            262.1144
#> 172              115.5546           176.6865            259.5634
#> 231              110.3592           169.8613            251.6816
#> 274              116.8440           179.5426            261.4687
#> 318              109.7750           171.6038            249.8745
#> 366              115.2937           176.4884            259.9847
#> 423              118.4986           184.0894            263.2314
#> 481              118.2499           177.7636            264.0023
#> 547              118.9649           184.0913            263.8267
#> 607              112.0499           175.9282            253.0315
#> 683              115.8695           180.1018            258.7312
#> 734              120.8302           187.1707            266.3642
#> 789              112.6191           175.7111            254.0429
#> 833              108.9003           173.3949            248.1354
#> 888              108.8372           174.9930            246.5288
#> 933              122.9780           185.5711            271.2892
#> 975              113.6087           174.9555            256.3395
#> 1022             115.2036           177.2672            258.8240
#> 1074             107.2670           166.2807            246.6752
#> 1138             117.7201           177.8367            263.2601
#> 1184             108.4449           169.2951            247.6681
#> 1233             126.3190           187.2450            277.2585
#> 1280             107.9727           169.0229            247.6734
#> 1325             115.1839           175.3195            259.6588
#> 1377             117.1196           177.0666            263.4115
#> 1423             118.4356           181.9341            263.5311
#> 1494             118.1130           179.3648            263.6328
#>      sd_metabolic_rate
#> 1             287.2098
#> 63            284.2527
#> 116           299.2248
#> 172           286.4324
#> 231           276.9689
#> 274           289.4335
#> 318           278.9577
#> 366           284.7535
#> 423           294.9590
#> 481           289.4706
#> 547           295.7750
#> 607           284.3443
#> 683           290.9656
#> 734           299.8616
#> 789           284.5617
#> 833           279.5519
#> 888           283.0143
#> 933           298.3323
#> 975           284.0090
#> 1022          286.7968
#> 1074          272.1432
#> 1138          288.7500
#> 1184          276.5339
#> 1233          301.0390
#> 1280          274.6089
#> 1325          284.3051
#> 1377          285.9191
#> 1423          293.1589
#> 1494          290.4994