Simulate body masses for a population

Draws body mass measurements for a population of birds (of all the same species) given the population size and either the species id or the mean and potentially standard deviation of body mass for that species.

simulate_population(
  species_abundance = NULL,
  species_mean = NULL,
  species_sd = NULL,
  species_code = NULL
)

Arguments

species_abundance
integer number of individuals to draw. Required.

species_mean
numeric, mean body mass (in grams) for this species.

species_sd
numeric, standard deviation of body mass for this species.

species_code
species ID for this species.

Value

a dataframe with species_abundance rows and columns for: species_code, species_mean, species_sd, species_abundance, simulation_method, and individual_mass.

Details

Fills in the necessary parameters based on the parameters provided and passes these to draw_population().

species_abundance and either species_mean or species_code must be provided. Depending on which parameters are provided:

species_mean and species_sd: If both mean and standard deviation body size are provided, uses the values provided to draw species_abundance individuals.

species_mean but no species_sd: If mean is provided but standard deviation is not, estimates the standard deviation based on the scaling relationship between mean and standard deviation of body mass (see estimate_sd()).

species_code but no species_mean or species_sd: If parameter values are not provided, but a species code is, look up the mean and standard deviation body mass for that species (see lookup_species_pars().