Define a species

Creates a list with taxonomic/identifying information and parameters for mean and standard deviation of body mass.

species_define(
  aou = NA,
  genus = NA,
  species = NA,
  mean_size = NA,
  sd_size = NA,
  sim_species_id = 1
)

Arguments

aou
AOU

genus
genus

species
species

mean_size
mean body size

sd_size
sd of body size

sim_species_id
identifier; if using taxonomic info, defaults to AOU. If not, defaults to 1. Supplying other values can be useful for simulation models.

Value

list with species parameter information

Details

The identifying information used depends on which parameters are provided, with the following order of preference: AOU > genus + species > user provided mean and sd > user provided mean and estimated sd.

Examples

species_define(aou = 2881)

#> $aou
#> [1] 2881
#> 
#> $genus
#> [1] "Perdix"
#> 
#> $species
#> [1] "perdix"
#> 
#> $mean_size
#> [1] 405.5
#> 
#> $sd_size
#> [1] 31
#> 
#> $sd_method
#> [1] "AOU lookup"
#> 
#> $sim_species_id
#> [1] 2881
#> 

species_define(genus = "Perdix", species = "perdix")

#> $aou
#> [1] 2881
#> 
#> $genus
#> [1] "Perdix"
#> 
#> $species
#> [1] "perdix"
#> 
#> $mean_size
#> [1] 405.5
#> 
#> $sd_size
#> [1] 31
#> 
#> $sd_method
#> [1] "Scientific name lookup"
#> 
#> $sim_species_id
#> [1] 2881
#> 

species_define(mean_size = 400, sd_size = 30)

#> $aou
#> [1] NA
#> 
#> $genus
#> [1] NA
#> 
#> $species
#> [1] NA
#> 
#> $mean_size
#> [1] 400
#> 
#> $sd_size
#> [1] 30
#> 
#> $sd_method
#> [1] "Mean and SD provided"
#> 
#> $sim_species_id
#> [1] 1
#> 

species_define(mean_size = 400)

#> $aou
#> [1] NA
#> 
#> $genus
#> [1] NA
#> 
#> $species
#> [1] NA
#> 
#> $mean_size
#> [1] 400
#> 
#> $sd_size
#> [1] 28.55916
#> 
#> $sd_method
#> [1] "SD estimated from mean"
#> 
#> $sim_species_id
#> [1] 1
#>