Estimate a species' standard deviation of body mass based on its mean body mass

Using the parameters estimated in get_sd_parameters, estimate the standard deviation of body mass for a species based only on its mean body mass.

species_estimate_sd(sp_mean, pars = NULL)

Arguments

sp_mean

mean body mass, in grams

pars

list containing $slope and $intercept, generated by get_sd_parameters(). If not provided, estimated by running get_sd_parameters(raw_masses).

Value

the estimated standard deviation of body mass