Simulate individual measurements for many populations

For a community (i.e. a collection of populations of different species, or of the same species at different points in time or locations, etc), simulate individual-level size and metabolic rate measurements.

community_generate(
  community_data_table,
  abundance_column_name = "speciestotal"
)

Arguments

community_data_table

dataframe containing at least one of aou, genus and species, or mean_size and a column for species abundances

abundance_column_name

character, the name of the column with species abundances. Defaults to "speciestotal".

Value

a dataframe one row per individual, all columns from community_data_table, and additional columns for species attributes.

Specifically:

aou: the AOU, if provided

sim_species_id: the sim_species_id if provided

genus: the genus associated with the AOU if provided, or the genus if provided

species: the species associated with the AOU if provided, or the species if provided

individual_mass: the simulated body mass (in grams) for this individual

individual_bmr: the simulated basal metabolic rate for this individual

mean_size: the mean body mass for this species (i.e. the parameter used for simulation)

sd_size: the standard deviation of body mass for this species (i.e. the parameter used for simulation)

abundance: the number of individuals simulated of this species (i.e. parameter used for simulation)

sd_method: the method for finding the standard deviation for body mass for this species

Examples

demo_community <- community_generate(demo_route_clean)
head(demo_community)
#>   record_id   routedataid countrynum statenum route rpid year count10 count20
#> 1    900000 9009911011994        900       99     1  101 1994       8      12
#> 2    900000 9009911011994        900       99     1  101 1994       8      12
#> 3    900000 9009911011994        900       99     1  101 1994       8      12
#> 4    900000 9009911011994        900       99     1  101 1994       8      12
#> 5    900000 9009911011994        900       99     1  101 1994       8      12
#> 6    900000 9009911011994        900       99     1  101 1994       8      12
#>   count30 count40 count50 stoptotal speciestotal  aou sim_species_id  genus
#> 1      15      12      15         5           62 4730           4730 Alauda
#> 2      15      12      15         5           62 4730           4730 Alauda
#> 3      15      12      15         5           62 4730           4730 Alauda
#> 4      15      12      15         5           62 4730           4730 Alauda
#> 5      15      12      15         5           62 4730           4730 Alauda
#> 6      15      12      15         5           62 4730           4730 Alauda
#>    species individual_mass individual_bmr mean_size  sd_size abundance
#> 1 arvensis        35.68001       134.2976    37.475 3.300613        62
#> 2 arvensis        41.82402       150.4062    37.475 3.300613        62
#> 3 arvensis        38.95494       142.9751    37.475 3.300613        62
#> 4 arvensis        30.57105       120.2861    37.475 3.300613        62
#> 5 arvensis        38.12620       140.7997    37.475 3.300613        62
#> 6 arvensis        35.92477       134.9538    37.475 3.300613        62
#>    sd_method
#> 1 AOU lookup
#> 2 AOU lookup
#> 3 AOU lookup
#> 4 AOU lookup
#> 5 AOU lookup
#> 6 AOU lookup