Compute grouped summary statistics for a community

Calculate summary statistics (total, mean, and standard deviation for abundance, biomass, and metabolic rate) for a community, with flexible grouping options.

community_summarize(
  community,
  level = c("year", "species", "species_and_year", "custom"),
  id_vars = NULL
)

Arguments

community

a dataframe with one row per individual, columns for individual_mass and individual_bmr, and any additional relevant grouping columns. For example, the output of community_generate.

level

a character string specifying whether to group according to "year", "species", "species_and_year", or "custom". To use "custom", specify the grouping variables using id_vars

id_vars

a vector of column names to group by (if using level = "custom")

Value

a dataframe with the total, mean, and standard deviation of body mass and metabolic rate summarized over the individuals in community, grouped according to the grouping variables.

Details

For a dataframe of individuals from either numerous populations (of different species, or of the same species at different points in time), calculate summary statistics for each population or timestep, or for a different set of custom-specified grouping variables.

Examples

bbs_route <- filter_bbs_survey(demo_route_raw)
community_data <- community_generate(bbs_route)
community_summarize(community_data)
#>        routedataid countrynum statenum route rpid year species_designator
#> 1    9009911011994        900       99     1  101 1994                AOU
#> 63   9009911011995        900       99     1  101 1995                AOU
#> 116  9009911011996        900       99     1  101 1996                AOU
#> 172  9009911011997        900       99     1  101 1997                AOU
#> 231  9009911011998        900       99     1  101 1998                AOU
#> 274  9009911011999        900       99     1  101 1999                AOU
#> 318  9009911012000        900       99     1  101 2000                AOU
#> 366  9009911012001        900       99     1  101 2001                AOU
#> 423  9009911012002        900       99     1  101 2002                AOU
#> 481  9009911012003        900       99     1  101 2003                AOU
#> 547  9009911012004        900       99     1  101 2004                AOU
#> 607  9009911012005        900       99     1  101 2005                AOU
#> 683  9009911012006        900       99     1  101 2006                AOU
#> 734  9009911012007        900       99     1  101 2007                AOU
#> 789  9009911012008        900       99     1  101 2008                AOU
#> 833  9009911012009        900       99     1  101 2009                AOU
#> 888  9009911012010        900       99     1  101 2010                AOU
#> 933  9009911012011        900       99     1  101 2011                AOU
#> 975  9009911012012        900       99     1  101 2012                AOU
#> 1022 9009911012013        900       99     1  101 2013                AOU
#> 1074 9009911012014        900       99     1  101 2014                AOU
#> 1138 9009911012015        900       99     1  101 2015                AOU
#> 1184 9009911012016        900       99     1  101 2016                AOU
#> 1233 9009911012017        900       99     1  101 2017                AOU
#> 1280 9009911012018        900       99     1  101 2018                AOU
#> 1325 9009911012019        900       99     1  101 2019                AOU
#> 1377 9009911012020        900       99     1  101 2020                AOU
#> 1423 9009911012021        900       99     1  101 2021                AOU
#> 1494 9009911012022        900       99     1  101 2022                AOU
#>      total_abundance total_biomass total_metabolic_rate total_richness
#> 1               1361      157003.4             352344.6             27
#> 63              1443      162259.4             366086.8             27
#> 116             1413      166877.9             369664.5             27
#> 172             1381      158999.5             357523.1             27
#> 231             1415      157590.2             357897.4             27
#> 274             1412      163510.4             367447.5             27
#> 318             1423      157603.5             357064.3             27
#> 366             1462      168701.5             379950.0             27
#> 423             1443      169619.9             377964.2             27
#> 481             1431      170094.9             378629.0             27
#> 547             1484      175413.7             390017.8             27
#> 607             1499      168710.1             380379.5             27
#> 683             1450      168200.8             375682.5             27
#> 734             1450      174394.2             385474.4             27
#> 789             1449      162471.3             366751.0             27
#> 833             1403      152805.2             348098.6             27
#> 888             1419      154343.9             349202.7             27
#> 933             1488      182288.3             402015.5             27
#> 975             1470      167006.5             376922.6             27
#> 1022            1443      167113.4             374837.4             27
#> 1074            1439      156470.7             358078.3             27
#> 1138            1374      162259.3             362141.1             27
#> 1184            1442      155837.6             356254.3             27
#> 1233            1430      179685.4             395188.0             27
#> 1280            1415      152846.8             350668.5             27
#> 1325            1465      168518.1             380111.4             27
#> 1377            1441      169635.4             380723.4             27
#> 1423            1437      170162.3             378397.5             27
#> 1494            1453      170938.5             381907.3             27
#>      mean_individual_mass sd_individual_mass mean_metabolic_rate
#> 1                115.3588           178.4827            258.8866
#> 63               112.4459           175.3540            253.6984
#> 116              118.1018           186.5550            261.6168
#> 172              115.1336           176.5018            258.8871
#> 231              111.3712           172.1708            252.9310
#> 274              115.8006           176.3661            260.2319
#> 318              110.7544           175.2302            250.9236
#> 366              115.3909           177.7823            259.8837
#> 423              117.5467           182.3794            261.9295
#> 481              118.8643           179.4472            264.5905
#> 547              118.2033           182.6107            262.8153
#> 607              112.5485           176.7614            253.7555
#> 683              116.0006           179.8658            259.0914
#> 734              120.2718           185.0072            265.8444
#> 789              112.1265           175.2904            253.1063
#> 833              108.9132           173.1846            248.1102
#> 888              108.7695           175.8953            246.0907
#> 933              122.5055           185.6088            270.1717
#> 975              113.6099           175.0435            256.4099
#> 1022             115.8097           178.7516            259.7626
#> 1074             108.7357           169.3229            248.8383
#> 1138             118.0926           179.3992            263.5671
#> 1184             108.0704           169.0132            247.0557
#> 1233             125.6541           185.7020            276.3552
#> 1280             108.0189           169.3626            247.8223
#> 1325             115.0294           174.5868            259.4617
#> 1377             117.7206           178.0205            264.2077
#> 1423             118.4149           182.6680            263.3247
#> 1494             117.6452           178.7454            262.8405
#>      sd_metabolic_rate
#> 1             287.9700
#> 63            284.3703
#> 116           297.6795
#> 172           285.9658
#> 231           279.8597
#> 274           286.0590
#> 318           282.6823
#> 366           285.9284
#> 423           292.7602
#> 481           291.5874
#> 547           293.9297
#> 607           285.4411
#> 683           290.7071
#> 734           297.5945
#> 789           284.1006
#> 833           279.5574
#> 888           284.0219
#> 933           298.5724
#> 975           283.9412
#> 1022          288.1862
#> 1074          275.5811
#> 1138          290.3735
#> 1184          275.9924
#> 1233          299.3674
#> 1280          274.6548
#> 1325          283.6725
#> 1377          287.2722
#> 1423          293.8106
#> 1494          289.8445