Compute grouped summary statistics for a community

Calculate summary statistics (total, mean, and standard deviation for abundance, biomass, and metabolic rate) for a community, with flexible grouping options.

community_summarize(
  community,
  level = c("year", "species", "species_and_year", "custom"),
  id_vars = NULL
)

Arguments

community

a dataframe with one row per individual, columns for individual_mass and individual_bmr, and any additional relevant grouping columns. For example, the output of community_generate.

level

a character string specifying whether to group according to "year", "species", "species_and_year", or "custom". To use "custom", specify the grouping variables using id_vars

id_vars

a vector of column names to group by (if using level = "custom")

Value

a dataframe with the total, mean, and standard deviation of body mass and metabolic rate summarized over the individuals in community, grouped according to the grouping variables.

Details

For a dataframe of individuals from either numerous populations (of different species, or of the same species at different points in time), calculate summary statistics for each population or timestep, or for a different set of custom-specified grouping variables.

Examples

bbs_route <- demo_route_raw %>% filter_bbs_survey()
community_data <- community_generate(bbs_route)
community_summarize(community_data)
#> # A tibble: 29 × 15
#>    routedataid count…¹ state…² route  rpid  year speci…³ total…⁴ total…⁵ total…⁶
#>    <chr>         <dbl>   <dbl> <dbl> <dbl> <int> <chr>     <int>   <dbl>   <dbl>
#>  1 9009911011…     900      99     1   101  1994 aou        1361 157437. 352736.
#>  2 9009911011…     900      99     1   101  1995 aou        1443 162013. 365755.
#>  3 9009911011…     900      99     1   101  1996 aou        1413 167124. 370405.
#>  4 9009911011…     900      99     1   101  1997 aou        1381 158851. 357227.
#>  5 9009911011…     900      99     1   101  1998 aou        1415 157390. 357778.
#>  6 9009911011…     900      99     1   101  1999 aou        1412 163685. 367699.
#>  7 9009911012…     900      99     1   101  2000 aou        1423 157097. 356342.
#>  8 9009911012…     900      99     1   101  2001 aou        1462 168187. 379208.
#>  9 9009911012…     900      99     1   101  2002 aou        1443 171141. 380254.
#> 10 9009911012…     900      99     1   101  2003 aou        1431 168967. 377130.
#> # … with 19 more rows, 5 more variables: total_richness <int>,
#> #   mean_individual_mass <dbl>, sd_individual_mass <dbl>,
#> #   mean_metabolic_rate <dbl>, sd_metabolic_rate <dbl>, and abbreviated
#> #   variable names ¹​countrynum, ²​statenum, ³​species_designator,
#> #   ⁴​total_abundance, ⁵​total_biomass, ⁶​total_metabolic_rate