Compute grouped summary statistics for a community

Calculate summary statistics (total, mean, and standard deviation for abundance, biomass, and metabolic rate) for a community, with flexible grouping options.

community_summarize(
  community,
  level = c("year", "species", "species_and_year", "custom"),
  id_vars = NULL
)

Arguments

community

a dataframe with one row per individual, columns for individual_mass and individual_bmr, and any additional relevant grouping columns. For example, the output of community_generate.

level

a character string specifying whether to group according to "year", "species", "species_and_year", or "custom". To use "custom", specify the grouping variables using id_vars

id_vars

a vector of column names to group by (if using level = "custom")

Value

a dataframe with the total, mean, and standard deviation of body mass and metabolic rate summarized over the individuals in community, grouped according to the grouping variables.

Details

For a dataframe of individuals from either numerous populations (of different species, or of the same species at different points in time), calculate summary statistics for each population or timestep, or for a different set of custom-specified grouping variables.

Examples

bbs_route <- filter_bbs_survey(demo_route_raw)
community_data <- community_generate(bbs_route)
community_summarize(community_data)
#>        routedataid countrynum statenum route rpid year species_designator
#> 1    9009911011994        900       99     1  101 1994                AOU
#> 63   9009911011995        900       99     1  101 1995                AOU
#> 116  9009911011996        900       99     1  101 1996                AOU
#> 172  9009911011997        900       99     1  101 1997                AOU
#> 231  9009911011998        900       99     1  101 1998                AOU
#> 274  9009911011999        900       99     1  101 1999                AOU
#> 318  9009911012000        900       99     1  101 2000                AOU
#> 366  9009911012001        900       99     1  101 2001                AOU
#> 423  9009911012002        900       99     1  101 2002                AOU
#> 481  9009911012003        900       99     1  101 2003                AOU
#> 547  9009911012004        900       99     1  101 2004                AOU
#> 607  9009911012005        900       99     1  101 2005                AOU
#> 683  9009911012006        900       99     1  101 2006                AOU
#> 734  9009911012007        900       99     1  101 2007                AOU
#> 789  9009911012008        900       99     1  101 2008                AOU
#> 833  9009911012009        900       99     1  101 2009                AOU
#> 888  9009911012010        900       99     1  101 2010                AOU
#> 933  9009911012011        900       99     1  101 2011                AOU
#> 975  9009911012012        900       99     1  101 2012                AOU
#> 1022 9009911012013        900       99     1  101 2013                AOU
#> 1074 9009911012014        900       99     1  101 2014                AOU
#> 1138 9009911012015        900       99     1  101 2015                AOU
#> 1184 9009911012016        900       99     1  101 2016                AOU
#> 1233 9009911012017        900       99     1  101 2017                AOU
#> 1280 9009911012018        900       99     1  101 2018                AOU
#> 1325 9009911012019        900       99     1  101 2019                AOU
#> 1377 9009911012020        900       99     1  101 2020                AOU
#> 1423 9009911012021        900       99     1  101 2021                AOU
#> 1494 9009911012022        900       99     1  101 2022                AOU
#>      total_abundance total_biomass total_metabolic_rate total_richness
#> 1               1361      157254.0             352670.6             27
#> 63              1443      161957.7             365697.1             27
#> 116             1413      166653.6             369683.2             27
#> 172             1381      159665.5             358335.9             27
#> 231             1415      157864.6             358633.0             27
#> 274             1412      163811.7             367736.8             27
#> 318             1423      157871.9             357445.4             27
#> 366             1462      168931.8             380234.6             27
#> 423             1443      171012.1             379746.9             27
#> 481             1431      169437.4             377771.2             27
#> 547             1484      176026.3             390620.7             27
#> 607             1499      168789.5             380461.1             27
#> 683             1450      168162.3             375725.9             27
#> 734             1450      174182.5             385050.5             27
#> 789             1449      162321.1             366974.2             27
#> 833             1403      153767.9             349652.2             27
#> 888             1419      153900.0             348733.1             27
#> 933             1488      182139.1             402565.8             27
#> 975             1470      168002.6             378067.8             27
#> 1022            1443      166638.9             374147.8             27
#> 1074            1439      157624.0             359625.5             27
#> 1138            1374      162640.6             362966.0             27
#> 1184            1442      156393.2             357378.7             27
#> 1233            1430      177820.7             392624.8             27
#> 1280            1415      152440.8             349974.2             27
#> 1325            1465      167098.4             378511.8             27
#> 1377            1441      169275.7             380126.8             27
#> 1423            1437      170839.5             379263.3             27
#> 1494            1453      171910.4             383702.9             27
#>      mean_individual_mass sd_individual_mass mean_metabolic_rate
#> 1                115.5430           178.7358            259.1261
#> 63               112.2368           174.8198            253.4284
#> 116              117.9431           185.7543            261.6300
#> 172              115.6159           177.8755            259.4757
#> 231              111.5651           171.8186            253.4509
#> 274              116.0139           177.5512            260.4368
#> 318              110.9430           174.9143            251.1914
#> 366              115.5484           178.0639            260.0784
#> 423              118.5115           184.3518            263.1649
#> 481              118.4049           178.2737            263.9911
#> 547              118.6161           183.8069            263.2215
#> 607              112.6014           177.2100            253.8099
#> 683              115.9740           179.7110            259.1213
#> 734              120.1259           185.5541            265.5520
#> 789              112.0229           174.4732            253.2603
#> 833              109.5994           174.3495            249.2176
#> 888              108.4567           175.7126            245.7597
#> 933              122.4053           184.0368            270.5415
#> 975              114.2875           176.9338            257.1890
#> 1022             115.4808           178.1734            259.2847
#> 1074             109.5372           170.6143            249.9135
#> 1138             118.3702           178.9425            264.1674
#> 1184             108.4558           169.0939            247.8354
#> 1233             124.3502           183.5217            274.5628
#> 1280             107.7320           168.8387            247.3316
#> 1325             114.0603           172.4807            258.3699
#> 1377             117.4710           177.8734            263.7937
#> 1423             118.8862           184.0513            263.9272
#> 1494             118.3141           178.8988            264.0763
#>      sd_metabolic_rate
#> 1             288.3569
#> 63            283.7995
#> 116           296.6327
#> 172           287.4485
#> 231           279.5436
#> 274           287.0968
#> 318           282.8163
#> 366           286.3205
#> 423           295.2715
#> 481           290.3337
#> 547           295.3973
#> 607           285.7423
#> 683           290.4885
#> 734           297.8288
#> 789           283.0077
#> 833           280.8673
#> 888           283.2122
#> 933           296.7345
#> 975           286.0992
#> 1022          287.4691
#> 1074          277.4668
#> 1138          290.1480
#> 1184          276.1579
#> 1233          296.5648
#> 1280          274.2055
#> 1325          280.8283
#> 1377          286.9964
#> 1423          295.1909
#> 1494          290.2280