Simulate body masses for a population

Draws body mass measurements for a population of birds (of all the same species) given the population size and either (1) the species id or (2) the mean and potentially standard deviation of body mass for that species.

pop_generate(
  abundance = NA,
  aou = NA,
  genus = NA,
  species = NA,
  mean_size = NA,
  sd_size = NA,
  sim_species_id = 1
)

Arguments

abundance
integer number of individuals to draw. Required.

aou
aou

genus
genus

species
species

mean_size
numeric, mean body mass (in grams) for this species.

sd_size
numeric, standard deviation of body mass for this species.

sim_species_id
defaults AOU or 1

Value

a dataframe with abundance rows and columns for species attributes.

Details

Fills in the necessary parameters based on the parameters provided and passes these to ind_draw().

Examples

pop_generate(abundance = 5, aou = 2881)

#>    aou sim_species_id  genus species individual_mass individual_bmr mean_size
#> 1 2881           2881 Perdix  perdix        376.7256       720.9335     405.5
#> 2 2881           2881 Perdix  perdix        364.2060       703.7684     405.5
#> 3 2881           2881 Perdix  perdix        364.9729       704.8247     405.5
#> 4 2881           2881 Perdix  perdix        384.4316       731.4175     405.5
#> 5 2881           2881 Perdix  perdix        411.2038       767.3827     405.5
#>   sd_size abundance  sd_method
#> 1      31         5 AOU lookup
#> 2      31         5 AOU lookup
#> 3      31         5 AOU lookup
#> 4      31         5 AOU lookup
#> 5      31         5 AOU lookup

pop_generate(abundance = 5, genus = "Selasphorus", species = "calliope")

#>    aou sim_species_id       genus  species individual_mass individual_bmr
#> 1 4360           4360 Selasphorus calliope        2.535091       20.38152
#> 2 4360           4360 Selasphorus calliope        2.550593       20.47031
#> 3 4360           4360 Selasphorus calliope        2.800381       21.88039
#> 4 4360           4360 Selasphorus calliope        2.264706       18.80671
#> 5 4360           4360 Selasphorus calliope        3.039852       23.19867
#>   mean_size   sd_size abundance              sd_method
#> 1      2.65 0.1818394         5 Scientific name lookup
#> 2      2.65 0.1818394         5 Scientific name lookup
#> 3      2.65 0.1818394         5 Scientific name lookup
#> 4      2.65 0.1818394         5 Scientific name lookup
#> 5      2.65 0.1818394         5 Scientific name lookup

pop_generate(abundance = 5, mean_size = 20, sd_size = 3)

#>   aou sim_species_id genus species individual_mass individual_bmr mean_size
#> 1  NA              1  <NA>    <NA>        24.59786      103.01467        20
#> 2  NA              1  <NA>    <NA>        17.53184       80.91674        20
#> 3  NA              1  <NA>    <NA>        21.02924       92.12182        20
#> 4  NA              1  <NA>    <NA>        18.44234       83.89114        20
#> 5  NA              1  <NA>    <NA>        18.55212       84.24688        20
#>   sd_size abundance            sd_method
#> 1       3         5 Mean and SD provided
#> 2       3         5 Mean and SD provided
#> 3       3         5 Mean and SD provided
#> 4       3         5 Mean and SD provided
#> 5       3         5 Mean and SD provided