Simulate body masses for a population

Draws body mass measurements for a population of birds (of all the same species) given the population size and either (1) the species id or (2) the mean and potentially standard deviation of body mass for that species.

pop_generate(
  abundance = NULL,
  aou = NULL,
  genus = NULL,
  species = NULL,
  mean_size = NULL,
  sd_size = NULL,
  id = 1
)

Arguments

abundance

integer number of individuals to draw. Required.

aou

aou

genus

genus

species

species

mean_size

numeric, mean body mass (in grams) for this species.

sd_size

numeric, standard deviation of body mass for this species.

id

defaults AOU or 1

Value

a dataframe with abundance rows and columns for species attributes.

Details

Fills in the necessary parameters based on the parameters provided and passes these to ind_draw().

Examples

pop_generate(abundance = 5, aou = 2881)
#>    aou   id  genus species individual_mass mean_size sd_size abundance
#> 1 2881 2881 Perdix  perdix        429.0552     405.5      31         5
#> 2 2881 2881 Perdix  perdix        343.4386     405.5      31         5
#> 3 2881 2881 Perdix  perdix        412.7139     405.5      31         5
#> 4 2881 2881 Perdix  perdix        358.3712     405.5      31         5
#> 5 2881 2881 Perdix  perdix        411.9100     405.5      31         5
#>    sd_method
#> 1 AOU lookup
#> 2 AOU lookup
#> 3 AOU lookup
#> 4 AOU lookup
#> 5 AOU lookup
pop_generate(abundance = 5, genus = "Selasphorus", species = "calliope")
#>    aou   id       genus  species individual_mass mean_size   sd_size abundance
#> 1 4360 4360 Selasphorus calliope        2.897407      2.65 0.1828933         5
#> 2 4360 4360 Selasphorus calliope        2.529929      2.65 0.1828933         5
#> 3 4360 4360 Selasphorus calliope        2.756850      2.65 0.1828933         5
#> 4 4360 4360 Selasphorus calliope        2.861865      2.65 0.1828933         5
#> 5 4360 4360 Selasphorus calliope        2.885999      2.65 0.1828933         5
#>                sd_method
#> 1 Scientific name lookup
#> 2 Scientific name lookup
#> 3 Scientific name lookup
#> 4 Scientific name lookup
#> 5 Scientific name lookup
pop_generate(abundance = 5, mean_size = 20, sd_size = 3)
#>   aou id genus species individual_mass mean_size sd_size abundance
#> 1  NA  1  <NA>    <NA>        18.87004        20       3         5
#> 2  NA  1  <NA>    <NA>        16.59803        20       3         5
#> 3  NA  1  <NA>    <NA>        21.72198        20       3         5
#> 4  NA  1  <NA>    <NA>        21.95725        20       3         5
#> 5  NA  1  <NA>    <NA>        21.86808        20       3         5
#>              sd_method
#> 1 Mean and SD provided
#> 2 Mean and SD provided
#> 3 Mean and SD provided
#> 4 Mean and SD provided
#> 5 Mean and SD provided