Simulate body masses for a population

Draws body mass measurements for a population of birds (of all the same species) given the population size and either (1) the species id or (2) the mean and potentially standard deviation of body mass for that species.

pop_generate(
  abundance = NA,
  aou = NA,
  genus = NA,
  species = NA,
  mean_size = NA,
  sd_size = NA,
  sim_species_id = 1
)

Arguments

abundance

integer number of individuals to draw. Required.

aou

aou

genus

genus

species

species

mean_size

numeric, mean body mass (in grams) for this species.

sd_size

numeric, standard deviation of body mass for this species.

sim_species_id

defaults AOU or 1

Value

a dataframe with abundance rows and columns for species attributes.

Details

Fills in the necessary parameters based on the parameters provided and passes these to ind_draw().

Examples

pop_generate(abundance = 5, aou = 2881)
#>    aou sim_species_id  genus species individual_mass individual_bmr mean_size
#> 1 2881           2881 Perdix  perdix        429.6242       791.7381     405.5
#> 2 2881           2881 Perdix  perdix        415.1164       772.5818     405.5
#> 3 2881           2881 Perdix  perdix        399.3253       751.5110     405.5
#> 4 2881           2881 Perdix  perdix        469.1792       843.0510     405.5
#> 5 2881           2881 Perdix  perdix        422.1626       781.9092     405.5
#>   sd_size abundance  sd_method
#> 1      31         5 AOU lookup
#> 2      31         5 AOU lookup
#> 3      31         5 AOU lookup
#> 4      31         5 AOU lookup
#> 5      31         5 AOU lookup
pop_generate(abundance = 5, genus = "Selasphorus", species = "calliope")
#>    aou sim_species_id       genus  species individual_mass individual_bmr
#> 1 4360           4360 Selasphorus calliope        2.743092       21.56030
#> 2 4360           4360 Selasphorus calliope        3.211207       24.12369
#> 3 4360           4360 Selasphorus calliope        2.727166       21.47097
#> 4 4360           4360 Selasphorus calliope        2.727009       21.47009
#> 5 4360           4360 Selasphorus calliope        2.674164       21.17261
#>   mean_size   sd_size abundance              sd_method
#> 1      2.65 0.1818394         5 Scientific name lookup
#> 2      2.65 0.1818394         5 Scientific name lookup
#> 3      2.65 0.1818394         5 Scientific name lookup
#> 4      2.65 0.1818394         5 Scientific name lookup
#> 5      2.65 0.1818394         5 Scientific name lookup
pop_generate(abundance = 5, mean_size = 20, sd_size = 3)
#>   aou sim_species_id genus species individual_mass individual_bmr mean_size
#> 1  NA              1  <NA>    <NA>        21.56141       93.77802        20
#> 2  NA              1  <NA>    <NA>        15.62013       74.52233        20
#> 3  NA              1  <NA>    <NA>        18.58134       84.34148        20
#> 4  NA              1  <NA>    <NA>        17.40600       80.50219        20
#> 5  NA              1  <NA>    <NA>        23.95086      101.07535        20
#>   sd_size abundance            sd_method
#> 1       3         5 Mean and SD provided
#> 2       3         5 Mean and SD provided
#> 3       3         5 Mean and SD provided
#> 4       3         5 Mean and SD provided
#> 5       3         5 Mean and SD provided