Simulate body masses for a population

Draws body mass measurements for a population of birds (of all the same species) given the population size and either (1) the species AOU or (2) the mean and potentially standard deviation of body mass for that species.

pop_generate(
  abundance = NA,
  aou = NA,
  genus = NA,
  species = NA,
  mean_size = NA,
  sd_size = NA,
  sim_species_id = 1
)

Arguments

abundance

integer number of individuals to draw. Required.

aou

aou

genus

genus

species

species

mean_size

numeric, mean body mass (in grams) for this species.

sd_size

numeric, standard deviation of body mass for this species.

sim_species_id

defaults AOU or 1

Value

a dataframe with abundance rows - one record per individual - and columns for species attributes.

Specifically:

aou: the AOU, if provided

sim_species_id: the sim_species_id if provided

genus: the genus associated with the AOU if provided, or the genus if provided

species: the species associated with the AOU if provided, or the species if provided

individual_mass: the simulated body mass (in grams) for this individual

individual_bmr: the simulated basal metabolic rate for this individual

mean_size: the mean body mass for this species (i.e. the parameter used for simulation)

sd_size: the standard deviation of body mass for this species (i.e. the parameter used for simulation)

abundance: the number of individuals simulated of this species (i.e. parameter used for simulation)

sd_method: the method for finding the standard deviation for body mass for this species

Details

abundance is required, as well as one of: aou, genus and species, or mean_size.

Examples

pop_generate(abundance = 5, aou = 2881)
#>    aou sim_species_id  genus species individual_mass individual_bmr mean_size
#> 1 2881           2881 Perdix  perdix        429.9191       792.1255     405.5
#> 2 2881           2881 Perdix  perdix        413.0375       769.8211     405.5
#> 3 2881           2881 Perdix  perdix        429.2611       791.2610     405.5
#> 4 2881           2881 Perdix  perdix        396.6598       747.9309     405.5
#> 5 2881           2881 Perdix  perdix        370.4187       712.3073     405.5
#>   sd_size abundance  sd_method
#> 1      31         5 AOU lookup
#> 2      31         5 AOU lookup
#> 3      31         5 AOU lookup
#> 4      31         5 AOU lookup
#> 5      31         5 AOU lookup
pop_generate(abundance = 5, genus = "Selasphorus", species = "calliope")
#>    aou sim_species_id       genus  species individual_mass individual_bmr
#> 1 4360           4360 Selasphorus calliope        2.344318       19.27574
#> 2 4360           4360 Selasphorus calliope        2.613924       20.83144
#> 3 4360           4360 Selasphorus calliope        2.654867       21.06356
#> 4 4360           4360 Selasphorus calliope        2.361132       19.37421
#> 5 4360           4360 Selasphorus calliope        2.722592       21.44529
#>   mean_size   sd_size abundance              sd_method
#> 1      2.65 0.1818394         5 Scientific name lookup
#> 2      2.65 0.1818394         5 Scientific name lookup
#> 3      2.65 0.1818394         5 Scientific name lookup
#> 4      2.65 0.1818394         5 Scientific name lookup
#> 5      2.65 0.1818394         5 Scientific name lookup
pop_generate(abundance = 5, mean_size = 20, sd_size = 3)
#>   aou sim_species_id genus species individual_mass individual_bmr mean_size
#> 1  NA              1  <NA>    <NA>        21.69260       94.18452        20
#> 2  NA              1  <NA>    <NA>        18.49290       84.05507        20
#> 3  NA              1  <NA>    <NA>        22.99784       98.19111        20
#> 4  NA              1  <NA>    <NA>        24.31219      102.16022        20
#> 5  NA              1  <NA>    <NA>        22.53838       96.78836        20
#>   sd_size abundance            sd_method
#> 1       3         5 Mean and SD provided
#> 2       3         5 Mean and SD provided
#> 3       3         5 Mean and SD provided
#> 4       3         5 Mean and SD provided
#> 5       3         5 Mean and SD provided